Update ASM2d naming conventions
