+++ watertap/property_models/activated_sludge/tests/test_modified_asm2d_thermo.py
@@ -10,12 +10,12 @@
 # "https://github.com/watertap-org/watertap/"
 #################################################################################
 """
-Tests for ASM2d thermo property package.
-Authors: Chenyu Wang
+Tests for modified ASM2d thermo property package.
+Author: Marcus Holly
 """
 
 import pytest
-from pyomo.environ import ConcreteModel, Param, units, value, Var
+from pyomo.environ import ConcreteModel, Param, value, Var
 from pyomo.util.check_units import assert_units_consistent
 from idaes.core import MaterialBalanceType, EnergyBalanceType, MaterialFlowBasis
 
@@ -40,9 +40,7 @@ class TestParamBlock(object):
     @pytest.fixture(scope="class")
     def model(self):
         model = ConcreteModel()
-        model.params = ModifiedASM2dParameterBlock(
-            additional_solute_list=["S_K", "S_Mg"]
-        )
+        model.params = ModifiedASM2dParameterBlock()
 
         return model
 
@@ -54,7 +52,7 @@ class TestParamBlock(object):
         for i in model.params.phase_list:
             assert i == "Liq"
 
-        assert len(model.params.component_list) == 22
+        assert len(model.params.component_list) == 19
         for i in model.params.component_list:
             assert i in [
                 "H2O",
@@ -66,19 +64,16 @@ class TestParamBlock(object):
                 "S_NO3",
                 "S_O2",
                 "S_PO4",
-                "S_ALK",
+                "S_K",
+                "S_Mg",
+                "S_IC",
                 "X_AUT",
                 "X_H",
                 "X_I",
-                "X_MeOH",
-                "X_MeP",
                 "X_PAO",
                 "X_PHA",
                 "X_PP",
                 "X_S",
-                "X_TSS",
-                "S_K",
-                "S_Mg",
             ]
 
         assert isinstance(model.params.cp_mass, Param)
@@ -98,9 +93,7 @@ class TestStateBlock(object):
     @pytest.fixture(scope="class")
     def model(self):
         model = ConcreteModel()
-        model.params = ModifiedASM2dParameterBlock(
-            additional_solute_list=["S_K", "S_Mg"]
-        )
+        model.params = ModifiedASM2dParameterBlock()
 
         model.props = model.params.build_state_block([1])
 
@@ -117,12 +110,9 @@ class TestStateBlock(object):
         assert isinstance(model.props[1].temperature, Var)
         assert value(model.props[1].temperature) == 298.15
 
-        assert isinstance(model.props[1].alkalinity, Var)
-        assert value(model.props[1].alkalinity) == 1
-
         assert isinstance(model.props[1].conc_mass_comp, Var)
         # H2O should not appear in conc_mass_comp
-        assert len(model.props[1].conc_mass_comp) == 20
+        assert len(model.props[1].conc_mass_comp) == 18
         for i in model.props[1].conc_mass_comp:
             assert i in [
                 "S_A",
@@ -133,18 +123,16 @@ class TestStateBlock(object):
                 "S_NO3",
                 "S_O2",
                 "S_PO4",
+                "S_K",
+                "S_Mg",
+                "S_IC",
                 "X_AUT",
                 "X_H",
                 "X_I",
-                "X_MeOH",
-                "X_MeP",
                 "X_PAO",
                 "X_PHA",
                 "X_PP",
                 "X_S",
-                "X_TSS",
-                "S_K",
-                "S_Mg",
             ]
             assert value(model.props[1].conc_mass_comp[i]) == 0.1
 
@@ -155,12 +143,6 @@ class TestStateBlock(object):
                 if j == "H2O":
                     assert str(model.props[1].get_material_flow_terms(p, j)) == str(
                         model.props[1].flow_vol * model.props[1].params.dens_mass
-                    )
-                elif j == "S_ALK":
-                    assert str(model.props[1].get_material_flow_terms(p, j)) == str(
-                        model.props[1].flow_vol
-                        * model.props[1].alkalinity
-                        * (61 * units.kg / units.kmol)
                     )
                 else:
                     assert str(model.props[1].get_material_flow_terms(p, j)) == str(
@@ -184,10 +166,6 @@ class TestStateBlock(object):
                 if j == "H2O":
                     assert str(model.props[1].get_material_density_terms(p, j)) == str(
                         model.props[1].params.dens_mass
-                    )
-                elif j == "S_ALK":
-                    assert str(model.props[1].get_material_density_terms(p, j)) == str(
-                        model.props[1].alkalinity * (61 * units.kg / units.kmol)
                     )
                 else:
                     assert str(model.props[1].get_material_density_terms(p, j)) == str(
@@ -225,11 +203,10 @@ class TestStateBlock(object):
     def test_define_state_vars(self, model):
         sv = model.props[1].define_state_vars()
 
-        assert len(sv) == 5
+        assert len(sv) == 4
         for i in sv:
             assert i in [
                 "flow_vol",
-                "alkalinity",
                 "conc_mass_comp",
                 "temperature",
                 "pressure",
@@ -239,11 +216,10 @@ class TestStateBlock(object):
     def test_define_port_members(self, model):
         sv = model.props[1].define_state_vars()
 
-        assert len(sv) == 5
+        assert len(sv) == 4
         for i in sv:
             assert i in [
                 "flow_vol",
-                "alkalinity",
                 "conc_mass_comp",
                 "temperature",
                 "pressure",
@@ -253,11 +229,10 @@ class TestStateBlock(object):
     def test_define_display_vars(self, model):
         sv = model.props[1].define_display_vars()
 
-        assert len(sv) == 5
+        assert len(sv) == 4
         for i in sv:
             assert i in [
                 "Volumetric Flowrate",
-                "Molar Alkalinity",
                 "Mass Concentration",
                 "Temperature",
                 "Pressure",
