+++ watertap/property_models/activated_sludge/tests/test_simple_modified_asm2d_thermo.py
@@ -10,18 +10,18 @@
 # "https://github.com/watertap-org/watertap/"
 #################################################################################
 """
-Tests for modified ASM2d thermo property package.
-Author: Marcus Holly
+Tests for ASM2d thermo property package.
+Authors: Chenyu Wang
 """
 
 import pytest
-from pyomo.environ import ConcreteModel, Param, value, Var
+from pyomo.environ import ConcreteModel, Param, units, value, Var
 from pyomo.util.check_units import assert_units_consistent
 from idaes.core import MaterialBalanceType, EnergyBalanceType, MaterialFlowBasis
 
-from watertap.property_models.activated_sludge.new_modified_asm2d_properties import (
-    NewASM2dParameterBlock,
-    NewASM2dStateBlock,
+from watertap.property_models.activated_sludge.simple_modified_asm2d_properties import (
+    SimpleModifiedASM2dParameterBlock,
+    SimpleModifiedASM2dStateBlock,
 )
 from idaes.core.util.model_statistics import (
     fixed_variables_set,
@@ -40,19 +40,21 @@ class TestParamBlock(object):
     @pytest.fixture(scope="class")
     def model(self):
         model = ConcreteModel()
-        model.params = NewASM2dParameterBlock()
+        model.params = SimpleModifiedASM2dParameterBlock(
+            additional_solute_list=["S_K", "S_Mg"]
+        )
 
         return model
 
     @pytest.mark.unit
     def test_build(self, model):
-        assert model.params.state_block_class is NewASM2dStateBlock
+        assert model.params.state_block_class is SimpleModifiedASM2dStateBlock
 
         assert len(model.params.phase_list) == 1
         for i in model.params.phase_list:
             assert i == "Liq"
 
-        assert len(model.params.component_list) == 19
+        assert len(model.params.component_list) == 22
         for i in model.params.component_list:
             assert i in [
                 "H2O",
@@ -64,16 +66,19 @@ class TestParamBlock(object):
                 "S_NO3",
                 "S_O2",
                 "S_PO4",
-                "S_K",
-                "S_Mg",
-                "S_IC",
+                "S_ALK",
                 "X_AUT",
                 "X_H",
                 "X_I",
+                "X_MeOH",
+                "X_MeP",
                 "X_PAO",
                 "X_PHA",
                 "X_PP",
                 "X_S",
+                "X_TSS",
+                "S_K",
+                "S_Mg",
             ]
 
         assert isinstance(model.params.cp_mass, Param)
@@ -93,7 +98,9 @@ class TestStateBlock(object):
     @pytest.fixture(scope="class")
     def model(self):
         model = ConcreteModel()
-        model.params = NewASM2dParameterBlock()
+        model.params = SimpleModifiedASM2dParameterBlock(
+            additional_solute_list=["S_K", "S_Mg"]
+        )
 
         model.props = model.params.build_state_block([1])
 
@@ -110,9 +117,12 @@ class TestStateBlock(object):
         assert isinstance(model.props[1].temperature, Var)
         assert value(model.props[1].temperature) == 298.15
 
+        assert isinstance(model.props[1].alkalinity, Var)
+        assert value(model.props[1].alkalinity) == 1
+
         assert isinstance(model.props[1].conc_mass_comp, Var)
         # H2O should not appear in conc_mass_comp
-        assert len(model.props[1].conc_mass_comp) == 18
+        assert len(model.props[1].conc_mass_comp) == 20
         for i in model.props[1].conc_mass_comp:
             assert i in [
                 "S_A",
@@ -123,16 +133,18 @@ class TestStateBlock(object):
                 "S_NO3",
                 "S_O2",
                 "S_PO4",
-                "S_K",
-                "S_Mg",
-                "S_IC",
                 "X_AUT",
                 "X_H",
                 "X_I",
+                "X_MeOH",
+                "X_MeP",
                 "X_PAO",
                 "X_PHA",
                 "X_PP",
                 "X_S",
+                "X_TSS",
+                "S_K",
+                "S_Mg",
             ]
             assert value(model.props[1].conc_mass_comp[i]) == 0.1
 
@@ -143,6 +155,12 @@ class TestStateBlock(object):
                 if j == "H2O":
                     assert str(model.props[1].get_material_flow_terms(p, j)) == str(
                         model.props[1].flow_vol * model.props[1].params.dens_mass
+                    )
+                elif j == "S_ALK":
+                    assert str(model.props[1].get_material_flow_terms(p, j)) == str(
+                        model.props[1].flow_vol
+                        * model.props[1].alkalinity
+                        * (61 * units.kg / units.kmol)
                     )
                 else:
                     assert str(model.props[1].get_material_flow_terms(p, j)) == str(
@@ -166,6 +184,10 @@ class TestStateBlock(object):
                 if j == "H2O":
                     assert str(model.props[1].get_material_density_terms(p, j)) == str(
                         model.props[1].params.dens_mass
+                    )
+                elif j == "S_ALK":
+                    assert str(model.props[1].get_material_density_terms(p, j)) == str(
+                        model.props[1].alkalinity * (61 * units.kg / units.kmol)
                     )
                 else:
                     assert str(model.props[1].get_material_density_terms(p, j)) == str(
@@ -203,10 +225,11 @@ class TestStateBlock(object):
     def test_define_state_vars(self, model):
         sv = model.props[1].define_state_vars()
 
-        assert len(sv) == 4
+        assert len(sv) == 5
         for i in sv:
             assert i in [
                 "flow_vol",
+                "alkalinity",
                 "conc_mass_comp",
                 "temperature",
                 "pressure",
@@ -216,10 +239,11 @@ class TestStateBlock(object):
     def test_define_port_members(self, model):
         sv = model.props[1].define_state_vars()
 
-        assert len(sv) == 4
+        assert len(sv) == 5
         for i in sv:
             assert i in [
                 "flow_vol",
+                "alkalinity",
                 "conc_mass_comp",
                 "temperature",
                 "pressure",
@@ -229,10 +253,11 @@ class TestStateBlock(object):
     def test_define_display_vars(self, model):
         sv = model.props[1].define_display_vars()
 
-        assert len(sv) == 4
+        assert len(sv) == 5
         for i in sv:
             assert i in [
                 "Volumetric Flowrate",
+                "Molar Alkalinity",
                 "Mass Concentration",
                 "Temperature",
                 "Pressure",
