+++ watertap/unit_models/tests/test_electroNP_ZO.py
@@ -19,8 +19,8 @@ from pyomo.environ import (
 )
 from idaes.core import FlowsheetBlock
 from watertap.unit_models.electroNP_ZO import ElectroNPZO
-from watertap.property_models.activated_sludge.modified_asm2d_properties import (
-    ModifiedASM2dParameterBlock,
+from watertap.property_models.activated_sludge.simple_modified_asm2d_properties import (
+    SimpleModifiedASM2dParameterBlock,
 )
 from idaes.core.solvers import get_solver
 from idaes.core.util.model_statistics import degrees_of_freedom
@@ -41,7 +41,7 @@ class TestElectroNP:
         m = ConcreteModel()
         m.fs = FlowsheetBlock(dynamic=False)
 
-        m.fs.properties = ModifiedASM2dParameterBlock(
+        m.fs.properties = SimpleModifiedASM2dParameterBlock(
             additional_solute_list=["S_K", "S_Mg"]
         )
